% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{water_daily}
\alias{water_daily}
\title{Fetch daily values from USGS water data}
\usage{
water_daily(sites, params = "00060", start = NULL, end = NULL)
}
\arguments{
\item{sites}{Character vector of USGS site numbers}

\item{params}{Parameter codes. Default "00060" (discharge).}

\item{start}{Start date as "YYYY-MM-DD"}

\item{end}{End date as "YYYY-MM-DD"}
}
\value{
tibble: site, parameter, date, value, qualifier, site_name
}
\description{
Fetch daily values from USGS water data
}
