% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water.R
\name{water_realtime}
\alias{water_realtime}
\title{Fetch real-time (instantaneous) water data from USGS}
\usage{
water_realtime(sites, params = "00060", period = "P1D")
}
\arguments{
\item{sites}{Character vector of USGS site numbers (e.g., "01646500")}

\item{params}{Parameter codes: "00060" = discharge, "00065" = gage height,
"00010" = water temp. Default "00060".}

\item{period}{ISO 8601 duration (e.g., "P1D" = past 1 day, "P7D" = past week).
Default "P1D".}
}
\value{
tibble: site, parameter, datetime, value, qualifier, site_name
}
\description{
Fetch real-time (instantaneous) water data from USGS
}
